server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
redis==5.0.1
aiofiles==23.2.1
jinja2==3.1.2
# Removed packages that require Rust compilation
# langchain
# langchain-community
# python-jose[cryptography]
# passlib[bcrypt]
# slowapi